Added functionality to circumvent the lack of a JSON type on the PBF 1. When returning features from the server (caching to be added) assign the _id to be in properties, so it shows up when the feature is query in queryRederedFeatures
2. When the feature is queried the audits currently present in the state
are queried to find the one that matches and the feature is overwritten.
From here I need to use this data to return complex react component that
can show the demand summary table for each workbook. additionally could
pass in URL of parcel audit JPEG once that functionality has been
written in.